pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "maorch-eval"
version = "0.1.0"
description = "Multi-Agent Orchestration Evaluation Framework (MAO-Bench)"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "Apache-2.0" }
keywords = ["benchmark", "multi-agent", "llm", "orchestration", "evaluation"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Science/Research",
  "License :: OSI Approved :: Apache Software License",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
  "click>=8.1",
  "pydantic>=2.0",
  "rich>=13.0",
  "httpx>=0.27",
  "docker>=7.0",
  "pyyaml>=6.0",
  "jsonlines>=4.0",
  "opentelemetry-sdk>=1.25",
  "opentelemetry-exporter-otlp>=1.25",
  "anthropic>=0.40",
  "tiktoken>=0.7",
  "pandas>=2.0",
  "numpy>=1.26",
  "scipy>=1.13",
  "matplotlib>=3.9",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-asyncio>=0.23",
  "pytest-cov>=5.0",
  "ruff>=0.4",
  "mypy>=1.10",
  "black>=24.0",
  "types-pyyaml",
  "types-jsonlines",
  "pre-commit>=3.7",
]

[project.scripts]
maorch = "harness.cli:main"

[project.urls]
Homepage = "https://github.com/rachitpareek/multi-agent-orchestration-evals"
Repository = "https://github.com/rachitpareek/multi-agent-orchestration-evals"
Issues = "https://github.com/rachitpareek/multi-agent-orchestration-evals/issues"

[tool.hatch.build.targets.wheel]
packages = ["harness"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "C4", "UP", "SIM", "TCH"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "--cov=harness --cov-report=term-missing"

[tool.black]
line-length = 100
target-version = ["py311"]
